+++ package.json
@@ -12,5 +12,8 @@
     "@sveltejs/vite-plugin-svelte": "^2.4.2",
     "svelte": "^4.0.5",
     "vite": "^4.4.5"
+  },
+  "dependencies": {
+    "@supabase/supabase-js": "github:supabase/supabase-js"
   }
 }
