+++ src/lib/Login.svelte
@@ -1,0 +1,16 @@
+<script>
+
+</script>
+
+<main>
+
+    <div class="div-card">
+        This is the Login page
+    </div>
+</main>
+
+<style>
+    .div-card {
+        color: white;
+    }
+</style>
