+++ src/lib/Navbar.svelte
@@ -1,0 +1,19 @@
+<script>
+
+</script>
+
+<nav>
+    <ul>
+        <li><a href="#home">Home</a></li>
+        <li><a href="#login">Login</a></li>
+        <li><a href="#recipe">Recipe</a></li>
+    </ul>
+</nav>
+
+<style>
+    ul {
+        display: flex;
+        list-style-type: none;
+        gap: 1em;
+    }
+</style>
