+++ src/lib/stores.mjs
@@ -1,0 +1,3 @@
+import { writable } from "svelte/store";
+
+export const route = writable("#home");
